Rename Mainpage.js to MainPage.js
