+++ econet/src/MainComponent.js
@@ -1,7 +1,6 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { theme } from "./theme.js";
 import { ThemeProvider } from "@mui/material";
-import { RecoilRoot } from "recoil";
 
 // pages
 import MainPage from "./pages/MainPage.js";
@@ -17,41 +16,39 @@ import App from "./components/App";
 
 function MainComponent() {
   return (
-    <RecoilRoot>
-      <ThemeProvider theme={theme}>
-        <BrowserRouter>
-          <Routes>
-            <Route path="/" element={<App />}>
-              <Route index element={<MainPage />} />
-              <Route path="signin">
-                <Route index element={<SignInPage />} />
-              </Route>
-              <Route path="signup">
-                <Route index element={<SignUpPage />} />
-              </Route>
-              <Route path="setting">
-                <Route index element={<SettingPage />} />
-              </Route>
-              <Route path="today">
-                <Route index element={<TodayPage />} />
-              </Route>
-              <Route path="location">
-                <Route index element={<LocationPage />} />
-              </Route>
-              <Route path="econews">
-                <Route index element={<EcoNewsPage />} />
-              </Route>
-              <Route path="mypage">
-                <Route index element={<MyPage />} />
-              </Route>
-              <Route path="account">
-                <Route index element={<AccountSettingPage />} />
-              </Route>
+    <ThemeProvider theme={theme}>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<App />}>
+            <Route index element={<MainPage />} />
+            <Route path="signin">
+              <Route index element={<SignInPage />} />
             </Route>
-          </Routes>
-        </BrowserRouter>
-      </ThemeProvider>
-    </RecoilRoot>
+            <Route path="signup">
+              <Route index element={<SignUpPage />} />
+            </Route>
+            <Route path="setting">
+              <Route index element={<SettingPage />} />
+            </Route>
+            <Route path="today">
+              <Route index element={<TodayPage />} />
+            </Route>
+            <Route path="location">
+              <Route index element={<LocationPage />} />
+            </Route>
+            <Route path="econews">
+              <Route index element={<EcoNewsPage />} />
+            </Route>
+            <Route path="mypage">
+              <Route index element={<MyPage />} />
+            </Route>
+            <Route path="account">
+              <Route index element={<AccountSettingPage />} />
+            </Route>
+          </Route>
+        </Routes>
+      </BrowserRouter>
+    </ThemeProvider>
   );
 }
