+++ econet/src/pages/SignInPage.js
@@ -5,6 +5,8 @@ import SignInput from "../components/SignInput";
 import SignButton from "../components/SignButton";
 import styles from "../styles/SignPage.module.css";
 import { signinUser } from "../api/signin";
+import { useSetRecoilState } from "recoil";
+import { isLoggedInState } from "../atom/atoms"; // Recoil 상태 가져오기'
 
 function SignInPage() {
   const [userInfo, setUserInfo] = useState({
@@ -26,6 +28,7 @@ function SignInPage() {
     userInfo.password.length >= 8;
 
   const navigate = useNavigate();
+  const setIsLoggedIn = useSetRecoilState(isLoggedInState); // Recoil 상태 설정 함수
 
   const goSignUpPage = () => {
     navigate("/signup");
@@ -37,6 +40,7 @@ function SignInPage() {
     signinUser(userInfo.email, userInfo.password)
       .then((data) => {
         if (data) {
+          setIsLoggedIn(true); // 로그인 성공 시 상태를 true로 설정
           navigate("/");
         } else {
           throw new Error(
